16비트로 바꿈
+세미콜론 두개 뭐야;;
+플레이 시작 수정
+오프셋1000넘으면 다시
+wait노트 추가

+++ Assets/Scripts/RoutineMaker/SaveInfo.cs
@@ -29,7 +29,7 @@ public class ActionData
     public string NoteType;
     public int Gym;
     public float beatsPerMinute;
-    public int WaitBeat;
+    public float WaitBeat;
     public float LongTime;
     public int Multi;
 }
